requirements.txt
torch
transformers
sagemaker
boto3
python-dotenv
datasets